moving cmd/validations to pkg/validations
